+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Closes down API
 - Updates tests (uses PhpSpec and Behat)
 - Major API change (BC break!)
+- Configuration moved to a different package
 
 ### Added